Enable tests for Tuple type
